Move PaymentOption inside of CustomerAdapter
